Improvements to iff_to_nats_trx.py
Have added several functions to iff.py and nats.py to support this feature. First, the function gets the departure and arrival airports, then checks if it has taken off. If it has not taken off, it creates the gate to runway path using the iff track data. Functions in nats are used to retrieve the closest nodes for a given lat/lon, the list of adjacent nodes to an already identified node, and getting an adjacent node that is closer to another named node (i.e., a runway).